partial refactors

+++ src/perft.h
@@ -14,6 +14,6 @@ struct Perft_Result_Type {
 
 void debug_perft(Position& position, Perft_Result_Type& res, Ply depth, Ply ply);
 
-long long fast_perft(Position& position, Ply depth, Ply ply);
+uint64_t fast_perft(Position& position, Ply depth, Ply ply);
 
-long long uci_perft(Position& position, Ply depth, Ply ply);
+uint64_t uci_perft(Position& position, Ply depth, Ply ply);
